pyproject.toml
[project]
name = "clerk-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "reflex>=0.7.14",
    "reflex-clerk-api>=1.0.1",
]
